move login with other auth components
